services/gamification/src/main/java/org/jolly/multiplication/gamification/game/badgeprocessors/SilverBadgeProcessor.java
package org.jolly.multiplication.gamification.game.badgeprocessors;

import org.jolly.multiplication.gamification.challenge.ChallengeSolvedDTO;
import org.jolly.multiplication.gamification.game.domain.BadgeType;
import org.jolly.multiplication.gamification.game.domain.ScoreCard;
import org.springframework.stereotype.Component;

import java.util.List;
import java.util.Optional;

/**
 * @author jolly
 */
@Component
public class SilverBadgeProcessor implements BadgeProcessor {
    @Override
    public Optional<BadgeType> process(int currentScore, List<ScoreCard> scoreCardList, ChallengeSolvedDTO challenge) {
        if (currentScore > 150) {
            return Optional.of(BadgeType.SILVER);
        }
        return Optional.empty();
    }

    @Override
    public BadgeType badgeType() {
        return BadgeType.SILVER;
    }
}
